backend/requirements.txt
django==4.2.13
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-axes==6.4.0
drf-spectacular==0.27.2
python-dotenv==1.0.1
psycopg2-binary==2.9.9
redis==5.0.4
celery==5.3.6
celery[redis]==5.3.6
django-celery-beat==2.6.0
gunicorn==22.0.0
whitenoise==6.6.0
groq==0.9.0
langdetect==1.0.9
requests==2.32.2
sentry-sdk==2.3.1



